requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
Automat==24.8.1
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
cymem==2.0.10
defusedxml==0.7.1
distro==1.9.0
Django==5.0.6
django-cors-headers==4.5.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.55.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==10.4.0
preshed==3.0.9
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyDispatcher==2.0.7
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
razorpay==1.4.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.12.0
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
Twisted==24.10.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
w3lib==2.2.1
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.0
youtube-search-python==1.6.6
youtube-transcript-api==0.6.2
zope.interface==7.1.1
